pyproject.toml
[project]
name = "playwright-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "markdown>=3.10",
    "playwright>=1.55.0",
    "python-dotenv>=1.2.1",
    "weasyprint>=66.0",
]
